+++ R/GIS.R
@@ -1,121 +1,69 @@
-#' Set python path for `tiler`
+#' Determine the area of each zone in a raster
 #'
-#' By default, \pkg{tiler} is configured to use python 2, which may not be available on
-#' recent Linux distributions (e.g., Ubuntu 20.04+).
-#' Thus, explicitly set tiler options to find the correct python path on their system.
+#' Polygonize a raster and calculate the area of each polygon.
 #'
-#' @keywords internal
-.setTilerPythonPath <- function() {
-  pythonPth <- if (nzchar(Sys.which("python3"))) {
-    Sys.which("python3")
-  } else if (nzchar(Sys.which("python"))) {
-    Sys.which("python")
-  } else {
-    NULL
-  }
-
-  ## try a few other places -- OS-specific
-  if (is.null(pythonPth)) {
-    if (.isWindows()) {
-        pydir <- file.path(getOption("gdalUtils_gdalPath")[[1]]$path)
-        possPyBin <- file.path(pydir, "python.exe")
-        possPyBin3 <- file.path(pydir, "python3.exe")
-
-        pythonPth <- if (file.exists(possPyBin3)) {
-          possPyBin3
-        } else if (file.exists(possPyBin)) {
-          possPyBin
-        } else {
-          stop("Can't find python.exe or python3.exe")
-        }
-
-        out <- findOSGeo4W() # set path in tiler::tiler_options
-        if (length(out) == 0)
-          stop("Need to have OSGeo4W installed; see ?tiler")
-    } else {
-      stop("Can't find suitable python version in system path.")
-    }
-  }
-
-  tiler::tiler_options(python = pythonPth)
-  message("Setting tiler option `python = '", pythonPth, "'`.")
-}
-
-#' `areaAndPolyValue`
-#'
-#' Determine the area of each zone in a raster. TODO: improve description
-#'
-#' @param ras A `Raster*` object
+#' @param ras A `Raster` or `SpatRaster` object
+#' @param ... Additional arguments (not used)
 #'
 #' @return list containing: `sizeInHa`, the area; and `polyID`, the polygon ID.
 #'
 #' @export
-areaAndPolyValue <- function(ras) {
-  polyIndivSpecies <- gdal_polygonizeR(ras) # 99 seconds with full ras
+#' @rdname areaAndPolyValue
+areaAndPolyValue <- function(ras, ...) {
+  UseMethod("areaAndPolyValue", ras)
+}
+
+#' @export
+#' @rdname areaAndPolyValue
+areaAndPolyValue.Raster <- function(ras, ...) {
+  polyIndivSpecies <- gdal_polygonizeR(ras)
   pArea <- as.numeric(sf::st_area(polyIndivSpecies) / 1e4)
   list(sizeInHa = pArea, polyID = polyIndivSpecies$DN)
 }
 
-#' Polygonize with GDAL
+#' @export
+#' @rdname areaAndPolyValue
+areaAndPolyValue.SpatRaster <- function(ras, ...) {
+  polyIndivSpecies <- gdal_polygonizeR(ras)
+  pArea <- as.numeric(sf::st_area(polyIndivSpecies) / 1e4)
+  list(sizeInHa = pArea, polyID = polyIndivSpecies$DN)
+}
+
+#' Polygonize a raster
 #'
-#' Based on
-#' <https://johnbaumgartner.wordpress.com/2012/07/26/getting-rasters-into-shape-from-r/>.
+#' @param x a `RasterLayer`, `SpatRaster`, or character giving the filepath to a raster.
 #'
-#' @param x TODO: description needed
-#' @param outshape TODO: description needed
-#' @param gdalformat TODO: description needed
-#' @param pypath TODO: description needed
-#' @param readpoly TODO: description needed
-#' @param quiet TODO: description needed
+#' @param outshape character giving the filepath for the output shapefile.
+#'
+#' @param gdalformat GDAL driver to use. See `terra::gdal(drivers=TRUE)`.
+#'
+#' @param readpoly logical indicating whether the polygons object should be returned
+#'        (this was previously using `gdal_polygonize` on disk and required reading the file)
+#'
+#' @param pypath,quiet deprecated. maintained for backwards compatibility only (not used).
+#'
+#' @return if `readpoly = TRUE` (default), a `SpatVector` object; otherwise, `NULL`.
 #'
 gdal_polygonizeR <- function(x, outshape = NULL, gdalformat = "ESRI Shapefile", # nolint
                              pypath = NULL, readpoly = TRUE, quiet = TRUE) {
-  .setTilerPythonPath()
-
-  if (isTRUE(readpoly)) requireNamespace("rgdal", quietly = TRUE)
-  if (is.null(pypath)) {
-    pypath <- Sys.which("gdal_polygonize.py")
-    if (!nzchar(pypath)) {
-      if (reproducible::.requireNamespace("gdalUtils")) {
-        gdalUtils::gdal_setInstallation()
-      }
-      o <- options()
-      pypath <- file.path(o$gdalUtils_gdalPath[[2]]$path, "gdal_polygonize.py")
-      if (!nzchar(pypath)) {
-        stop("Need gdal_polygonize.py")
-      }
-    }
-  }
-  if (!file.exists(pypath)) stop("Can't find gdal_polygonize.py on your system.")
-  #owd <- getwd()
-  #on.exit(setwd(owd))
-  #setwd(dirname(pypath))
   if (is.null(outshape)) {
     outshape <- tempfile(fileext = ".shp")
-  } else {
-    outshape <- sub("\\.shp$", "", outshape)
-    fExists <- file.exists(paste(outshape, c("shp", "shx", "dbf"), sep = "."))
-    if (any(fExists))
-      stop(sprintf("File already exists: %s",
-                   toString(paste(outshape, c("shp", "shx", "dbf"), sep = ".")[fExists])),
-           call. = FALSE)
   }
-  if (is(x, "Raster")) {
-    f <- tempfile(fileext = ".tif")
-    rastpath <- normalizePath(f, mustWork = FALSE)
-    writeRaster(x, rastpath, datatype = assessDataType(x))
-  } else if (is.character(x)) {
-    rastpath <- normalizePath(x)
-  } else {
-    stop("x must be a file path (character string), or a Raster object.")
+  if (is(x, "Raster") || is(x, "character")) {
+    x <- terra::rast(x)
   }
-  system2(tiler::tiler_options()[["python"]],
-          args = (sprintf('"%1$s" "%2$s" -f "%3$s" "%4$s"',
-                          pypath, rastpath, gdalformat, outshape)))
+  if (terra::is.factor(x)) {
+    x <- as.numeric(x)
+  }
+
+  p <- terra::as.polygons(x) |>
+    terra::disagg()  ## to get POLYGONS from MULTIPOLYGONS
+  names(p) <- "DN" ## for backwards compatibility
+
+  terra::writeVector(p, outshape, filetype = gdalformat)
+
   if (isTRUE(readpoly)) {
-    shp <- sf::read_sf(dsn = dirname(outshape), layer = basename(file_path_sans_ext(outshape)))
-    sf::st_bbox(shp, extent(x))
-    return(shp)
+    return(p)
   } else {
     return(NULL)
   }
@@ -123,41 +71,88 @@ gdal_polygonizeR <- function(x, outshape = NULL, gdalformat = "ESRI Shapefile", 
 
 #' `.rasterToMemory`
 #'
-#' @param x A `Raster*` object
-#' @param ... Additional arguments passed to `raster`
+#' @param x A `Raster` or `SpatRaster` object
+#'
+#' @param ... Additional arguments passed to raster read function
 #'
 #' @export
 #' @rdname rasterToMemory
 .rasterToMemory <- function(x, ...) {
-  r <- raster(x, ...)
-  r <- raster::setValues(r, raster::getValues(r))
-  return(r)
+  UseMethod(".rasterToMemory", x)
 }
 
-#' Fasterize with crop & spTransform first
+#' @export
+#' @rdname rasterToMemory
+.rasterToMemory.Raster <- function(x, ...) {
+  raster::raster(x, ...) |>
+    raster::setValues(raster::getValues(x))
+}
+
+#' @export
+#' @rdname rasterToMemory
+.rasterToMemory.SpatRaster <- function(x, ...) {
+  terra::rast(x, ...) |>
+    terra::setValues(terra::values(x))
+}
+
+#' Rasterize following crop and reproject
 #'
-#' @param emptyRaster An empty raster with res, crs, extent all
-#'        correct for to pass to `fasterize`
-#' @param polygonToFasterize passed to `fasterize`, but it
-#'        will be cropped first if
-#'        `extent(emptyRaster) < extent(polygonToFasterize)`
-#' @param field passed to `fasterize`
+#' A simple wrapper around [terra::rasterize()].
+#'
+#' @param emptyRaster An empty `RasterLayer` or`SpatRaster` to use as a template.
+#'
+#' @param polygonToFasterize an `sf` or `SpatVector` object, which will be cropped first
+#'        if `extent(emptyRaster) < extent(polygonToFasterize)`.
+#'
+#' @inheritParams terra::rasterize
+#'
+#' @return an object of the same class as `emptyRaster`
 #'
 #' @export
+#'
+#' @examples
+#'
+#' ## using sf + raster
+#' f1 <- system.file("external/lux.shp", package = "raster")
+#' v1 <- sf::st_read(f1)
+#' r1 <- raster::raster(v1, ncols = 75, nrows = 100)
+#' raster::crs(r1) <- "epsg:4326"
+#' z1 <- fasterize2(r1, v1, "NAME_2")
+#'
+#' ## using terra
+#' f2 <- system.file("ex/lux.shp", package = "terra")
+#' v2 <- terra::vect(f2)
+#' r2 <- terra::rast(v2, ncols = 75, nrows = 100)
+#' z2 <- fasterize2(r2, v2, "NAME_2")
+#'
+#' terra::compareGeom(terra::rast(z1), z2)
 fasterize2 <- function(emptyRaster, polygonToFasterize, field) {
-  ras <- raster(emptyRaster)
-  if (extent(polygonToFasterize) > extent(ras)) {
-    polygonToFasterize <- Cache(cropInputs, polygonToFasterize, rasterToMatch = ras)
-  }
-  thePoly <- projectInputs(polygonToFasterize, targetCRS = crs(ras))
-  thePoly$polygonNum <- seq_along(thePoly)
-  if (!is.factor(thePoly[[field]])) {
-    thePoly[[field]] <- factor(thePoly[[field]])
-  }
-  aa2 <- fasterize::fasterize(sf::st_as_sf(thePoly), ras, field = field)
-  levels(aa2) <- data.frame(ID = seq_along(thePoly[[field]]),
-                            Factor = as.character(thePoly[[field]]),
-                            as.data.frame(thePoly))
+  asRaster <- ifelse(is(emptyRaster, "RasterLayer"), TRUE, FALSE)
 
-  aa2
+  if (!is(polygonToFasterize, "SpatVector")) {
+    polygonToFasterize <- terra::vect(polygonToFasterize)
+  }
+
+  ras <- terra::rast(emptyRaster)
+  if (terra::ext(polygonToFasterize) > terra::ext(ras)) {
+    polygonToFasterize <- reproducible::Cache(cropInputs, polygonToFasterize, rasterToMatch = ras)
+  }
+  thePoly <- reproducible::projectInputs(polygonToFasterize, targetCRS = raster::crs(ras))
+  if (!is(thePoly, "SpatVector")) {
+    thePoly <- terra::vect(thePoly)
+  }
+  thePoly$polygonNum <- seq_along(thePoly)
+  if (!is.factor(thePoly[[field]][[1]])) {
+    thePoly[[field]][[1]] <- factor(thePoly[[field]][[1]])
+  }
+  aa2 <- terra::rasterize(thePoly, ras, field = field)
+  # levels(aa2) <- data.frame(ID = seq_along(thePoly[[field]][[1]]),
+  #                           category = as.character(thePoly[[field]][[1]]),
+  #                           as.data.frame(thePoly))
+
+  if (isTRUE(asRaster)) {
+    return(raster::raster(aa2))
+  } else {
+    return(aa2)
+  }
 }