+++ NEWS.md
@@ -1,6 +1,13 @@
 # map (development version)
 
+* improved documentation
+
+# map v1.1.0
+
 * drop support for R 4.0 due to changes in dependency packages;
+* remove retiring dependency packages `sp`, `rgeos`, `rgdal` in favour of `terra` + `sf`;
+* remove dependency package `gdalUtils` in favour of `sf` + `terra`;
+* use most recent `tiler` version (0.3.0);
 * use `digest::digest(x, algo = "spooky")` instead of `fastdigest` (which was archived on CRAN);
 * remove unused dependencies `knitr` and `rmarkdown`;